refactor: extract LLM runtime into @pmk/llm
The provider layer (resolver, anthropic-api, claude-agent, config) lived
inside packages/cli, so the desktop app had to depend on the whole CLI
just to talk to a model. Move it to a standalone @pmk/llm package.

- packages/cli/src/llm/* and src/config.ts become thin re-export shims,
  keeping every existing import path working (554 -> 59 lines).
- Desktop drops @pmk/cli, @pmk/core and @pmk/rag for @pmk/llm alone;
  packages/rag drops its unused @pmk/core and @pmk/shared deps.
- @pmk/shared gains an exports map with a "browser" condition, which
  removes the hand-written Vite alias in electron.vite.config.ts. That
  alias reached across the monorepo by relative path and its own comment
  flagged it as brittle; resolution now goes through the package
  boundary as that comment proposed.

Also adds packages/core/src/index.js, the public entry point its
package.json "main" already pointed at.

LLM behaviour is unchanged: the existing suites (llm-anthropic-api,
llm-claude-agent, resolver) exercise the moved implementation through
the shims.
