+++ apps/desktop/package.json
@@ -21,9 +21,7 @@
     "@electron-toolkit/preload": "^3.0.1",
     "@electron-toolkit/utils": "^4.0.0",
     "@monaco-editor/react": "^4.7.0",
-    "@pmk/cli": "*",
-    "@pmk/core": "*",
-    "@pmk/rag": "*",
+    "@pmk/llm": "*",
     "@pmk/shared": "*",
     "gray-matter": "^4.0.3",
     "highlight.js": "^11.11.1",
@@ -41,7 +39,7 @@
     "@vitejs/plugin-react": "^4.3.4",
     "autoprefixer": "^10.4.20",
     "electron": "^41.0.0",
-    "electron-builder": "^26.8.1",
+    "electron-builder": "^26.15.3",
     "electron-vite": "^3.1.0",
     "postcss": "^8.4.49",
     "react": "^19.0.0",
