+++ packages/core/src/index.js
@@ -1,0 +1,5 @@
+/** Public package entry point for reusable core utilities. */
+module.exports = {
+  traceability: require("./traceability"),
+  confluenceSync: require("./confluence-sync"),
+};
